modulate the components out of Sponsors component
previously, all of the components were inside of the same file.
this isn't ideal for readability, so it is modulated to several
different files.
